🚚 rename hoc files to kebab-case
