+++ packages/kitchn/src/components/Badge/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 
 import { capitalize, isString } from "../..";
-import { withDecorator } from "../../hoc/withDecorator";
+import { withDecorator } from "../../hoc/with-decorator";
 import { KitchnComponent, NormalSizes } from "../../types";
 
 type Props = {
