+++ packages/kitchn/src/components/Collapse/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { RiArrowDownSLine } from "react-icons/ri";
 import styled from "styled-components";
 
-import { withDecorator } from "../../hoc/withDecorator";
+import { withDecorator } from "../../hoc/with-decorator";
 import { KitchnComponent } from "../../types";
 import Icon from "../Icon";
