+++ packages/kitchn/src/components/List/Unordered/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import { withDecorator } from "../../../hoc/withDecorator";
+import { withDecorator } from "../../../hoc/with-decorator";
 import { KitchnComponent } from "../../../types";
 import ListItem from "../Item";
