+++ packages/kitchn/src/components/Scroller/index.tsx
@@ -1,7 +1,7 @@
 import React, { useRef, useEffect, useState } from "react";
 import styled from "styled-components";
 
-import { withDecorator } from "../../hoc/withDecorator";
+import { withDecorator } from "../../hoc/with-decorator";
 import { KitchnComponent } from "../../types";
 import { convertRGBToRGBA } from "../../utils/convert-rgb-to-rgba";
