+++ packages/kitchn/src/components/Spacer/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import { withDecorator } from "../../hoc/withDecorator";
+import { withDecorator } from "../../hoc/with-decorator";
 import { KitchnComponent } from "../../types";
 import { Gap } from "../../types/theme";
 import { isNumber } from "../../utils/is-number";
