+++ packages/kitchn/src/hoc/index.ts
@@ -1,3 +1,3 @@
-export * from "./withBox";
-export * from "./withDecorator";
-export * from "./withScale";
+export * from "./with-box";
+export * from "./with-decorator";
+export * from "./with-scale";
