+++ packages/kitchn/src/native/hoc/index.ts
@@ -1,1 +1,1 @@
-export * from "./withScale";
+export * from "./with-scale";
