+++ packages/kitchn/src/native/hoc/with-scale.tsx
