+++ packages/kitchn/src/next/index.ts
@@ -1,3 +1,3 @@
 export * from "./components";
 
-export { default as withKitchnConfig } from "./hoc/withKitchnConfig";
+export { default as withKitchnConfig } from "./hoc/with-kitchn-config";
